Exits build if chroot fails to apt-get install
Exits the build script if the main `apt-get install` command in the chroot
fails. Also added an exit code check for the second set of chroot commands that
may be used in subsequent commits.

A typo in a single package name can cause the `apt-get install` command to
fail. Without checking the exit code, the build fails in a subsequent command
for a reason that will not be immediately clear to a developer.

+++ build.sh
@@ -90,6 +90,10 @@ cp /etc/resolv.conf chroot/etc/resolv.conf
 
 cp ../chroot.steps.part.1.sh ../chroot.steps.part.2.sh chroot
 chroot chroot/ /bin/bash /chroot.steps.part.1.sh
+if [[ $? -ne 0 ]]; then
+    echo "Error: Failed to execute chroot steps part 1."
+    exit 1
+fi
 
 # Copy the source FHS filesystem tree onto the build's chroot FHS tree, overwriting the base files where conflicts occur
 cd ..
@@ -137,6 +141,11 @@ done
 # Enter chroot again
 cd $BUILD_DIRECTORY
 chroot chroot/ /bin/bash /chroot.steps.part.2.sh
+if [[ $? -ne 0 ]]; then
+    echo "Error: Failed to execute chroot steps part 2."
+    exit 1
+fi
+
 rsync --archive chroot/var.cache.apt.archives/ ../$PKG_CACHE_DIRECTORY/$APT_PKG_CACHE_DIRECTORY
 rm -rf chroot/var.cache.apt.archives
 rsync --archive chroot/var.lib.apt.lists/ ../$PKG_CACHE_DIRECTORY/$APT_INDEX_CACHE_DIRECTORY
